ID continuation

+++ src/core/utility/functions.rs
@@ -12,21 +12,21 @@ pub fn ecs_pair(pred: u64, obj: u64) -> u64 {
 }
 
 #[inline(always)]
-pub fn ecs_pair_first(e: u64) -> u32 {
-    ecs_entity_t_hi(e & RUST_ECS_COMPONENT_MASK)
+pub fn ecs_pair_first(e: u64) -> u64 {
+    ecs_entity_t_hi(e & RUST_ECS_COMPONENT_MASK) as u64
 }
 
 #[inline(always)]
-pub fn ecs_pair_second(e: u64) -> u32 {
+pub fn ecs_pair_second(e: u64) -> u64 {
     ecs_entity_t_lo(e)
 }
 
 #[inline(always)]
-pub fn ecs_entity_t_lo(value: u64) -> u32 {
-    value as u32
+pub fn ecs_entity_t_lo(value: u64) -> u64 {
+    value as u32 as u64
 }
 
 #[inline(always)]
-pub fn ecs_entity_t_hi(value: u64) -> u32 {
-    (value >> 32) as u32
+pub fn ecs_entity_t_hi(value: u64) -> u64 {
+    value >> 32
 }
